[FIX] data type in examples

+++ examples/index.js
@@ -53,7 +53,7 @@ console.log( '\n' );
 
 // ----
 // Typed arrays...
-data = new Int32Array( 10 );
+data = new Float32Array( 10 );
 for ( i = 0; i < data.length; i++ ) {
 	data[ i ] = -2.5 * 0.5 * i;
 }
